requirements.txt
torchvision
numpy
qutip
matplotlib
tqdm